+++ cleveldb.go
@@ -1,3 +1,4 @@
+//go:build cleveldb
 // +build cleveldb
 
 package db