+++ cleveldb_iterator.go
@@ -1,3 +1,4 @@
+//go:build cleveldb
 // +build cleveldb
 
 package db
@@ -20,7 +21,7 @@ var _ Iterator = (*cLevelDBIterator)(nil)
 
 func newCLevelDBIterator(source *levigo.Iterator, start, end []byte, isReverse bool) *cLevelDBIterator {
 	if isReverse {
-		if end == nil || len(end) == 0 {
+		if len(end) == 0 {
 			source.SeekToLast()
 		} else {
 			source.Seek(end)
@@ -34,7 +35,7 @@ func newCLevelDBIterator(source *levigo.Iterator, start, end []byte, isReverse b
 			}
 		}
 	} else {
-		if start == nil || len(start) == 0 {
+		if len(start) == 0 {
 			source.SeekToFirst()
 		} else {
 			source.Seek(start)
@@ -56,7 +57,6 @@ func (itr cLevelDBIterator) Domain() ([]byte, []byte) {
 
 // Valid implements Iterator.
 func (itr cLevelDBIterator) Valid() bool {
-
 	// Once invalid, forever invalid.
 	if itr.isInvalid {
 		return false
@@ -64,28 +64,24 @@ func (itr cLevelDBIterator) Valid() bool {
 
 	// If source errors, invalid.
 	if itr.source.GetError() != nil {
-		itr.isInvalid = true
 		return false
 	}
 
 	// If source is invalid, invalid.
 	if !itr.source.Valid() {
-		itr.isInvalid = true
 		return false
 	}
 
 	// If key is end or past it, invalid.
-	var start = itr.start
-	var end = itr.end
-	var key = itr.source.Key()
+	start := itr.start
+	end := itr.end
+	key := itr.source.Key()
 	if itr.isReverse {
 		if start != nil && bytes.Compare(key, start) < 0 {
-			itr.isInvalid = true
 			return false
 		}
 	} else {
 		if end != nil && bytes.Compare(end, key) <= 0 {
-			itr.isInvalid = true
 			return false
 		}
 	}