+++ prefixdb_test.go
@@ -1,6 +1,12 @@
 package db
 
 import (
+	"bytes"
+	"encoding/binary"
+	"fmt"
+	"math/rand"
+	"path/filepath"
+	"sync"
 	"testing"
 
 	"github.com/stretchr/testify/require"
@@ -18,6 +24,109 @@ func mockDBWithStuff(t *testing.T) DB {
 	require.NoError(t, db.Set(bz("ke"), bz("valu")))
 	require.NoError(t, db.Set(bz("kee"), bz("valuu")))
 	return db
+}
+
+func taskKey(i, k int) []byte {
+	return []byte(fmt.Sprintf("task-%d-key-%d", i, k))
+}
+
+func randomValue() []byte {
+	b := make([]byte, 16)
+	rand.Read(b)
+	return b
+}
+
+func TestGolevelDB(t *testing.T) {
+	path := filepath.Join(t.TempDir(), "goleveldb")
+
+	db, err := NewGoLevelDB(path, "")
+	require.NoError(t, err)
+
+	Run(t, db)
+}
+
+/* We don't seem to test badger anywhere.
+func TestWithBadgerDB(t *testing.T) {
+	dir := t.TempDir()
+	path := filepath.Join(dir, "badgerdb")
+
+	db, err := NewBadgerDB(path, "")
+	require.NoError(t, err)
+
+	t.Run("BadgerDB", func(t *testing.T) { Run(t, db) })
+}
+*/
+
+func TestWithMemDB(t *testing.T) {
+	db := NewMemDB()
+
+	t.Run("MemDB", func(t *testing.T) { Run(t, db) })
+}
+
+// Run generates concurrent reads and writes to db so the race detector can
+// verify concurrent operations are properly synchronized.
+// The contents of db are garbage after Run returns.
+func Run(t *testing.T, db DB) {
+	t.Helper()
+
+	const numWorkers = 10
+	const numKeys = 64
+
+	var wg sync.WaitGroup
+	for i := 0; i < numWorkers; i++ {
+		wg.Add(1)
+		i := i
+		go func() {
+			defer wg.Done()
+
+			// Insert a bunch of keys with random data.
+			for k := 1; k <= numKeys; k++ {
+				key := taskKey(i, k) // say, "task-<i>-key-<k>"
+				value := randomValue()
+				if err := db.Set(key, value); err != nil {
+					t.Errorf("Task %d: db.Set(%q=%q) failed: %v",
+						i, string(key), string(value), err)
+				}
+			}
+
+			// Iterate over the database to make sure our keys are there.
+			it, err := db.Iterator(nil, nil)
+			if err != nil {
+				t.Errorf("Iterator[%d]: %v", i, err)
+				return
+			}
+			found := make(map[string][]byte)
+			mine := []byte(fmt.Sprintf("task-%d-", i))
+			for {
+				if key := it.Key(); bytes.HasPrefix(key, mine) {
+					found[string(key)] = it.Value()
+				}
+				it.Next()
+				if !it.Valid() {
+					break
+				}
+			}
+			if err := it.Error(); err != nil {
+				t.Errorf("Iterator[%d] reported error: %v", i, err)
+			}
+			if err := it.Close(); err != nil {
+				t.Errorf("Close iterator[%d]: %v", i, err)
+			}
+			if len(found) != numKeys {
+				t.Errorf("Task %d: found %d keys, wanted %d", i, len(found), numKeys)
+			}
+
+			// Delete all the keys we inserted.
+			for key := range mine {
+				bs := make([]byte, 4)
+				binary.LittleEndian.PutUint32(bs, uint32(key))
+				if err := db.Delete(bs); err != nil {
+					t.Errorf("Delete %q: %v", key, err)
+				}
+			}
+		}()
+	}
+	wg.Wait()
 }
 
 func TestPrefixDBSimple(t *testing.T) {