+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
+- remove mutex from prefixdb
+
 ## 0.6.7
 
 **2022-2-21**